refactor(scheduler): extract SearchMissing candidate enumeration
SearchMissing decided what to search for and then searched for it in the same two methods:
four QueryBuilders, a batch quality load, a cutoff filter and a season-pack grouping, all
interleaved with torznab calls and grabs. Under the plugin cut the enumeration reads core
tables the download plugin will not have, so it has to sit on the core side of the boundary.

AcquisitionCandidatesService answers three questions and does nothing else: which movies need
a search, which episodes need one, and how those episodes group into season packs. The
scheduler keeps every action — the searches, the pending-grab checks, the grabs, the progress
events.

The move is literal. Every SQL fragment, select list, filter and log string is byte-identical;
the only rewrites are the ones the new signatures force, chiefly the pack's weakest-quality
gate now reading each target's own files instead of a shared quality map.

Verified by golden-file comparison against the real database, not by tests: the enumeration
has none, and a silent error here means a media simply stops being searched.
